refactor(tools): small refactoring of tests package
